requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pydantic==1.10.13
PyYAML==6.0.1
requests==2.31.0
beautifulsoup4==4.12.2
httpx==0.24.1
